+++ lib/src/client.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:math' show min;
 import 'dart:typed_data' show Uint8List, BytesBuilder;
+import 'package:speed_test_dart/speed_test_dart.dart';
 import 'package:tus_client_dart/src/tus_client_base.dart';
 
 import 'exceptions.dart';
@@ -82,19 +83,57 @@ class TusClient extends TusClientBase {
     }
   }
 
+  Future<void> setUploadTestServers() async {
+    final tester = SpeedTestDart();
+
+    try {
+      final settings = await tester.getSettings();
+      final servers = settings.servers;
+
+      bestServers = await tester.getBestServers(
+        servers: servers,
+      );
+    } catch (_) {
+      bestServers = null;
+    }
+  }
+
+  Future<void> uploadSpeedTest() async {
+    final tester = SpeedTestDart();
+
+    // If bestServers are null or they are empty, we will not measure upload speed
+    // as it wouldn't be accurate at all
+    if (bestServers == null || (bestServers?.isEmpty ?? true)) {
+      uploadSpeed = null;
+      return;
+    }
+
+    try {
+      uploadSpeed = await tester.testUploadSpeed(servers: bestServers ?? []);
+    } catch (_) {
+      uploadSpeed = null;
+    }
+  }
+
   /// Start or resume an upload in chunks of [maxChunkSize] throwing
   /// [ProtocolException] on server error
   Future<void> upload({
     Function(double, Duration)? onProgress,
-    Function(TusClient)? onStart,
+    Function(TusClient, Duration?)? onStart,
     Function()? onComplete,
     required Uri uri,
     Map<String, String>? metadata = const {},
     Map<String, String>? headers = const {},
+    bool measureUploadSpeed = false,
   }) async {
     setUploadData(uri, headers, metadata);
 
     final isResumamble = await isResumable();
+
+    if (measureUploadSpeed) {
+      await setUploadTestServers();
+      await uploadSpeedTest();
+    }
 
     if (!isResumamble) {
       await createUpload();
@@ -113,7 +152,16 @@ class TusClient extends TusClientBase {
     final client = getHttpClient();
 
     if (onStart != null) {
-      onStart(this);
+      Duration? estimate;
+      if (uploadSpeed != null) {
+        final _workedUploadSpeed = uploadSpeed! * 1000000;
+
+        estimate = Duration(
+          seconds: (totalBytes / _workedUploadSpeed).round(),
+        );
+      }
+      // The time remaining to finish the upload
+      onStart(this, estimate);
     }
 
     while (!_pauseUpload && _offset < totalBytes) {
@@ -138,17 +186,23 @@ class TusClient extends TusClientBase {
             if (onProgress != null && !_pauseUpload) {
               // Total byte sent
               final totalSent = _offset + maxChunkSize;
+              double _workedUploadSpeed = 1.0;
 
-              // The total upload speed in bytes/ms
-              final uploadSpeed =
-                  totalSent / uploadStopwatch.elapsedMilliseconds;
+              // If upload speed != null, it means it has been measured
+              if (uploadSpeed != null) {
+                // Multiplied by 10^6 to convert from Mb/s to b/s
+                _workedUploadSpeed = uploadSpeed! * 1000000;
+              } else {
+                _workedUploadSpeed =
+                    totalSent / uploadStopwatch.elapsedMilliseconds;
+              }
 
               // The data that hasn't been sent yet
               final remainData = totalBytes - totalSent;
 
               // The time remaining to finish the upload
               final estimate = Duration(
-                milliseconds: (remainData / uploadSpeed).round(),
+                seconds: (remainData / _workedUploadSpeed).round(),
               );
 
               final progress = totalSent / totalBytes * 100;