+++ lib/src/tus_client_base.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:cross_file/cross_file.dart';
+import 'package:speed_test_dart/classes/server.dart';
 import 'package:tus_client_dart/tus_client_dart.dart';
 
 abstract class TusClientBase {
@@ -15,6 +16,12 @@ abstract class TusClientBase {
 
   /// Any additional headers
   Map<String, String>? headers;
+
+  /// Upload speed in Mb/s
+  double? uploadSpeed;
+
+  /// List of [Server] that are good for testing speed
+  List<Server>? bestServers;
 
   TusClientBase(
     this.file, {
@@ -31,7 +38,7 @@ abstract class TusClientBase {
   /// Starts an upload
   Future<void> upload({
     Function(double, Duration)? onProgress,
-    Function(TusClient)? onStart,
+    Function(TusClient, Duration?)? onStart,
     Function()? onComplete,
     required Uri uri,
     Map<String, String>? metadata = const {},
@@ -51,6 +58,12 @@ abstract class TusClientBase {
   String? generateFingerprint() {
     return file.path.replaceAll(RegExp(r"\W+"), '.');
   }
+
+  /// Sets to servers to test for upload speed
+  Future<void> setUploadTestServers();
+
+  /// Measures the upload speed of the device
+  Future<void> uploadSpeedTest();
 
   /// Override this to customize creating 'Upload-Metadata'
   String generateMetadata() {